Service/AC: add ac:i service
